Remove Entities and Move Entity to Models Foulder
