--- Refhub_Ir/Data/Mapping/BookRelationConfiguration.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Refhub_Ir.Models.Entities;
-
-namespace Refhub_Ir.Mapping
-{
-    public class BookRelationConfiguration : IEntityTypeConfiguration<BookRelation>
-    {
-        public void Configure(EntityTypeBuilder<BookRelation> builder)
-        {
-            builder.HasKey(rb => new { rb.BookId, rb.RelatedBookId });
-
-            builder.HasOne(br=>br.Book).WithMany(b=>b.RelatedTo)
-                .HasForeignKey(b=>b.BookId)
-                .OnDelete(DeleteBehavior.Restrict);
-
-            builder.HasOne(br=>br.RelatedBook).WithMany(b=>b.RelatedFrom)
-                .HasForeignKey(br=>br.RelatedBookId)
-                .OnDelete(DeleteBehavior.Restrict);
-        }
-    }
-}
