--- Refhub_Ir/Data/Mapping/CategoryConfiguration.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Refhub_Ir.Models.Entities;
-
-namespace Refhub_Ir.Mapping
-{
-    public class CategoryConfiguration : IEntityTypeConfiguration<Category>
-    {
-        public void Configure(EntityTypeBuilder<Category> builder)
-        {
-            builder.HasKey(c => c.Id);
-            builder.Property(c => c.Name).IsRequired().HasMaxLength(50);
-            builder.Property(c => c.slug).IsRequired();
-            builder.Property(c => c.Description);
-
-            builder.HasMany(c => c.Books).WithOne(c => c.Category)
-                .HasForeignKey(c => c.CategoryId)
-                .OnDelete(DeleteBehavior.Cascade);
-        }
-    }
-}
