+++ Refhub_Ir/Models/BookKeyword.cs
@@ -1,4 +1,4 @@
-﻿namespace Refhub_Ir.Models.Entities
+﻿namespace Refhub_Ir.Models
 {
     public class BookKeyword
     {
