+++ Refhub_Ir/Models/BookRelation.cs
@@ -1,4 +1,4 @@
-﻿namespace Refhub_Ir.Models.Entities
+﻿namespace Refhub_Ir.Models
 {
     public class BookRelation
     {
